requirements.txt
pytesseract
Pillow
openai
google-generativeai
anthropic